requirements.txt
psycopg2
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy
pandas
python-dotenv
nltk
wordcloud
matplotlib
streamlit